xử lý bị lách ck 09, bán online, thu cũ đổi mới

+++ src/app/sales-management/page/sale-online/hh-web/sale-online.component.ts
@@ -32,6 +32,7 @@ import { environment } from '@environments/environment';
 import { Package } from '@app/sales-management/model/ticket/common-model/package.model';
 import { PromotionSelectComponent } from '@app/sales-management/component/promotion/promotion-select.component';
 import { DomSanitizer } from '@angular/platform-browser';
+import { DialogConfirmComponent } from '@app/_components/dialog/dialog-confirm/dialog-confirm.component';
 
 const { DISCOUNT_LIST,
     GUARANTEE_LIST,
@@ -227,6 +228,23 @@ export class SaleOnlineComponent implements OnInit, AfterViewInit {
 
     // #region customer
     handleAddCustomer(customer: Customer) {
+        const ma_kh_old = this.ticket.masterInfo.ma_kh ? this.ticket.masterInfo.ma_kh : '';
+        const ma_kh_new = customer ? customer.ma_kh : '';
+        const msg_confirm_change = 'Có thay đổi mã khách, hệ thống sẽ tự động xóa các chương trình chiết khấu đã áp dụng trên phiếu. Xác nhận thực hiện?';
+        const style_css = 'font-size:16px;';
+        if (ma_kh_old !== '' && ma_kh_old !== ma_kh_new) {
+            this.commonService.openDialog(DialogConfirmComponent, { title: msg_confirm_change, style_css: style_css })
+                .afterClosed().subscribe(result => {
+                    if (!result) {
+                        //không xác nhận => reset về mã cũ
+                        this.ticket.masterInfo.ma_kh = ma_kh_old;
+                        return;
+                    }
+                    // xóa và reset ck 09 khi xác nhận => reset
+                    this.resetDiscount09();
+                });
+        }
+
         this.commonService.focusControl2(this.tabIndex.nvvc);
         this.saleOnlineService.removeDiscountForCustomer();
         this.saleOnlineService.setInfoCustomer(customer);
@@ -257,18 +275,18 @@ export class SaleOnlineComponent implements OnInit, AfterViewInit {
         this.customerApiService.getOneById(ma_kh).subscribe(result => {
             if (result.success && result.result) {
                 const customer: any = result.result;
-                this.handleAddCustomer(customer);
 
                 this.ticketApiService.getRankCustomer({ ma_kh: customer.ma_kh }).subscribe(result => {
                     const { ma_hang, mau_chu, tl_tich_diem } = result.result as any;
-                    this.ticket.masterInfo.ma_hang = ma_hang;
-                    this.ticket.masterInfo.tl_tich_diem = tl_tich_diem;
+                    this.ticket.masterInfo.ma_hang = ma_hang || '';
+                    this.ticket.masterInfo.tl_tich_diem = tl_tich_diem || 0;
                     this.generateLabelWithColor(ma_hang, mau_chu);
+                    this.handleAddCustomer(customer);
+                    // mở dialog add khách hàng nhưng ở chế độ update
+                    this.addOrUpdateCustomer = 'update';
+                    this.openAddCustomerDialog(customer.ma_kh);
                 });
 
-                // mở dialog add khách hàng nhưng ở chế độ update
-                this.addOrUpdateCustomer = 'update';
-                this.openAddCustomerDialog(customer.ma_kh);
             } else {
                 this.commonService.showMessageByContent(Language.content.exists_customer_yn_no, ma_kh);
                 this.saleOnlineService.resetCustomerInfo(this.ticket);
@@ -283,18 +301,19 @@ export class SaleOnlineComponent implements OnInit, AfterViewInit {
             { keyword: '', componentName: SEARCH_COMPONENT_NAME.CUSTOMER, title: this.getLabel('tlt_customer_list') }, 'search-style-dialog')
             .afterClosed()
             .subscribe((customer: Customer) => {
-                customer && this.handleAddCustomer(customer)
 
                 customer && this.ticketApiService.getRankCustomer({ ma_kh: customer.ma_kh }).subscribe(result => {
                     const { ma_hang, mau_chu, tl_tich_diem } = result.result as any;
-                    this.ticket.masterInfo.ma_hang = ma_hang;
-                    this.ticket.masterInfo.tl_tich_diem = tl_tich_diem;
+                    this.ticket.masterInfo.ma_hang = ma_hang || '';
+                    this.ticket.masterInfo.tl_tich_diem = tl_tich_diem || 0;
                     this.generateLabelWithColor(ma_hang, mau_chu);
+                    this.handleAddCustomer(customer);
+                    // mở dialog add khách hàng nhưng ở chế độ update
+                    this.addOrUpdateCustomer = 'update';
+                    this.openAddCustomerDialog(customer.ma_kh);
                 });
 
-                // mở dialog add khách hàng nhưng ở chế độ update
-                this.addOrUpdateCustomer = 'update';
-                this.openAddCustomerDialog(customer.ma_kh);
+
             });
     }
 
@@ -499,6 +518,7 @@ export class SaleOnlineComponent implements OnInit, AfterViewInit {
             this.handleCheckDeposit(merchandise.ma_vt, false);
             this.saleOnlineService.setIsNeedCalcDiscount(true);
             this.handleRemoveDiscountProgram(merchandise.ma_imei);
+            this.removeDiscount09(merchandise.ma_imei);
         }
     }
     handleRemoveDiscountProgram(ma_imei: string) {
@@ -798,23 +818,33 @@ export class SaleOnlineComponent implements OnInit, AfterViewInit {
     * Xử lý ck 09
     */
     applyDiscount09ForMerchandise(merchandiseResponse: any) {
-        const ma_kh = this.ticket.masterInfo.ma_kh.trim();
-        const ma_hang = this.ticket.masterInfo.ma_hang.trim();
+        const ma_kh = this.ticket.masterInfo.ma_kh ? this.ticket.masterInfo.ma_kh.trim() : '';
+        const ma_hang = this.ticket.masterInfo.ma_hang ? this.ticket.masterInfo.ma_hang.trim() : '';
         const ngay_ct = new Date(this.ticket.masterInfo.ngay_ct);
-        const ma_imei = merchandiseResponse.ma_imei.trim();
-        const ma_vt = merchandiseResponse.ma_vt.trim();
+        const ma_imei = merchandiseResponse.ma_imei ? merchandiseResponse.ma_imei.trim() : '';
+        const ma_vt = merchandiseResponse.ma_vt ? merchandiseResponse.ma_vt.trim() : '';
 
-        this.imeiApiService.getDiscountRankCustomer(ma_kh, ma_hang, ngay_ct, ma_imei, ma_vt, TICKET_CODE.RETAIL).subscribe((res: any) => {
+        this.imeiApiService.getDiscountRankCustomer(ma_kh, ma_hang, ngay_ct, ma_imei, ma_vt, TICKET_CODE.ONLINE).subscribe((res: any) => {
             if (res.success && res.result) {
                 const discount = res.result[0] as any;
                 // add vào tab ck
                 if (discount) {
                     this.discountService.addNew([discount], this.ticket.discount);
                 }
-                // tính lại tiền
+                // tính lại tiền khi có ck
+                this.saleOnlineService.calcMoney();
+            } else {
+                // tính lại tiền khi ko có ck
                 this.saleOnlineService.calcMoney();
             }
         });
+    }
+
+    removeDiscount09(ma_imei: string) {
+        this.ticket.discount = this.ticket.discount.filter(item => item.ma_imei.trim().toUpperCase() !== ma_imei.trim().toUpperCase());
+
+        // tính lại tiền
+        this.saleOnlineService.calcMoney();
     }
 
     /*
@@ -842,6 +872,20 @@ export class SaleOnlineComponent implements OnInit, AfterViewInit {
             )` : 'Mã khách';
 
         this.ma_kh_label = this.sanitizer.bypassSecurityTrustHtml(html) as string;
+    }
+
+    resetDiscount09() {
+        this.ticket.discount = this.ticket.discount.filter(item => item.loai_ck !== DISCOUNT_TYPE.DISCOUNT_CUSTOMER_RANK);
+
+        this.ticket.merchandise.map(item => {
+            this.applyDiscount09ForMerchandise(item);
+
+            item.tl_ck09 = 0;
+            item.tien_kb09 = 0;
+            item.tien_max09 = 0;
+            item.tien_ck09 = 0;
+            item.tl_ck_sau_vat09 = 0;
+        })
     }
 }
